Refactor matching
The previous system for matching was pointlessly abstracted to the implementation. This system is slightly better, but real world use will be the primary determination of effectiveness.
